Improved probes for faster startup time (#237)
The primary thing here is to remove the initialDelaySeconds on readiness and
liveness; since we have a startup probe, these serve no purpoes. Also, we
switch the startup probe to look for readiness so that the service is known
to be ready on startup (unlike the readiness probe, startup probe will kill
the pod).